src/backend/requirements.txt
asgiref==3.4.1
Django==3.1.12
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.12.3
pytz==2021.3
sqlparse==0.2.4
tzdata==2021.5
